integration tests: Move find_examples_dir() to testlib
This allows to use it for multiple tests.
